switched to gwtgl typed arrays

+++ src/com/google/gwt/corp/emul/java/nio/DirectReadOnlyShortBufferAdapter.java
@@ -18,8 +18,8 @@ package java.nio;
 
 import com.googlecode.gwtquake.client.HasArrayBufferView;
 
-import com.google.gwt.typedarrays.client.ArrayBufferView;
-import com.google.gwt.typedarrays.client.Int16Array;
+import com.googlecode.gwtgl.array.ArrayBufferView;
+import com.googlecode.gwtgl.array.Int16Array;
 
 /**
  * This class wraps a byte buffer to be a short buffer.
@@ -81,7 +81,7 @@ final class DirectReadOnlyShortBufferAdapter extends ShortBuffer implements HasA
 //        if (position == limit) {
 //            throw new BufferUnderflowException();
 //        }
-        return shortArray.get(position++);
+        return (short) shortArray.get(position++);
     }
 
     @Override
@@ -89,7 +89,7 @@ final class DirectReadOnlyShortBufferAdapter extends ShortBuffer implements HasA
 //        if (index < 0 || index >= limit) {
 //            throw new IndexOutOfBoundsException();
 //        }
-        return shortArray.get(index);
+        return (short) shortArray.get(index);
     }
 
     @Override
